+++ test/inputs/collection_select_input_test.rb
@@ -172,6 +172,12 @@ class CollectionSelectInputTest < ActionView::TestCase
     assert_select 'select[required]'
   end
 
+  test 'collection input with select type should generate required html attribute only with blank option or prompt' do
+    with_input_for @user, :name, :select, prompt: 'Name...', collection: ['Jose', 'Carlos']
+    assert_select 'select.required'
+    assert_select 'select[required]'
+  end
+
   test 'collection input with select type should not generate required html attribute without blank option' do
     with_input_for @user, :name, :select, :include_blank => false, :collection => ['Jose' , 'Carlos']
     assert_select 'select.required'